results output and css changes
reduced amount of search results to 5. reset the array at the beginning of the movieSearch eventListener. This seems to have fixed the bug that was not allowing multiple searches in a row. Also changed the CSS to allow for minimum screen size for everything to fit correctly. Media query will be next for CSS.